add memory cached

+++ main.py
@@ -97,6 +97,11 @@ def train_val_pipeline(MODEL_NAME, DATASET_NAME, params, net_params, dirs):
             val_loader = DataLoader(valset, batch_size=params['batch_size'], shuffle=False, drop_last=drop_last, collate_fn=dataset.collate)
             test_loader = DataLoader(testset, batch_size=params['batch_size'], shuffle=False, drop_last=drop_last, collate_fn=dataset.collate)
 
+            ckpt_dir = os.path.join(root_ckpt_dir, "RUN_" + str(split_number))
+            if not os.path.exists(ckpt_dir):
+                os.makedirs(ckpt_dir)
+            torch.save(model.state_dict(), '{}.pkl'.format(ckpt_dir + "/epoch_init"))
+
             with tqdm(range(params['epochs'])) as t:
                 for epoch in t:
                     t.set_description('Epoch %d' % epoch)    
@@ -119,26 +124,26 @@ def train_val_pipeline(MODEL_NAME, DATASET_NAME, params, net_params, dirs):
 
                     _, epoch_test_acc = evaluate_network(model, device, test_loader, epoch)
                     t.set_postfix(time=time.time()-start, lr=optimizer.param_groups[0]['lr'],
-                                  memory="{:.2%}".format(torch.cuda.max_memory_allocated(device=device)/total_memory),
+                                  memory="{:.4%}".format(torch.cuda.max_memory_cached(device=device)/total_memory),
                                   train_loss=epoch_train_loss, val_loss=epoch_val_loss,
                                   train_acc=epoch_train_acc, val_acc=epoch_val_acc,
                                   test_acc=epoch_test_acc)  
 
                     per_epoch_time.append(time.time()-start)
-                    per_epoch_memory.append(torch.cuda.max_memory_allocated(device=device))
+                    per_epoch_memory.append(torch.cuda.max_memory_cached(device=device))
 
                     # Saving checkpoint
-                    ckpt_dir = os.path.join(root_ckpt_dir, "RUN_" + str(split_number))
-                    if not os.path.exists(ckpt_dir):
-                        os.makedirs(ckpt_dir)
                     torch.save(model.state_dict(), '{}.pkl'.format(ckpt_dir + "/epoch_" + str(epoch)))
 
                     files = glob.glob(ckpt_dir + '/*.pkl')
                     for file in files:
                         epoch_nb = file.split('_')[-1]
-                        epoch_nb = int(epoch_nb.split('.')[0])
-                        if epoch_nb < epoch-1:
-                            os.remove(file)
+                        try:
+                            epoch_nb = int(epoch_nb.split('.')[0])
+                            if epoch_nb < epoch-1 and epoch_nb % 50 != 0:
+                                os.remove(file)
+                        except:
+                            pass
 
                     scheduler.step(epoch_val_loss)
 
@@ -171,6 +176,7 @@ def train_val_pipeline(MODEL_NAME, DATASET_NAME, params, net_params, dirs):
 
     print("TOTAL TIME TAKEN: {:.4f}hrs".format((time.time()-t0)/3600))
     print("AVG TIME PER EPOCH: {:.4f}s".format(np.mean(per_epoch_time)))
+    print("AVG MEMORY PER EPOCH: {:.4%}".format(np.mean(per_epoch_time))/total_memory)
 
     # Final test accuracy value averaged over 10-fold
     print("""\n\n\nFINAL RESULTS\n\nTEST ACCURACY averaged: {:.4f} with s.d. {:.4f}"""          .format(np.mean(np.array(avg_test_acc))*100, np.std(avg_test_acc)*100))
@@ -186,7 +192,7 @@ def train_val_pipeline(MODEL_NAME, DATASET_NAME, params, net_params, dirs):
     with open(write_file_name + '.txt', 'w') as f:
         f.write("""Dataset: {},\nModel: {}\n\nparams={}\n\nnet_params={}\n\n{}\n\nTotal Parameters: {}\n\n
     FINAL RESULTS\nTEST ACCURACY averaged: {:.4f} with s.d. {:.4f}\nTRAIN ACCURACY averaged: {:.4f} with s.d. {:.4f}\n\n
-    Total Time Taken: {:.4f} hrs\n Percentage of Average Memory taken per Epoch: {:.2%} \nAverage Time Per Epoch: {:.4f} s\n\n\nAll Splits Test Accuracies: {}"""
+    Total Time Taken: {:.4f} hrs\n Percentage of Average Memory taken per Epoch: {:.4%} \nAverage Time Per Epoch: {:.4f} s\n\n\nAll Splits Test Accuracies: {}"""
           .format(DATASET_NAME, MODEL_NAME, params, net_params, model, net_params['total_param'],
                   np.mean(np.array(avg_test_acc))*100, np.std(avg_test_acc)*100,
                   np.mean(np.array(avg_train_acc))*100, np.std(avg_train_acc)*100,
